requirements.txt
fastapi==0.104.1
uvicorn==0.21.0
python-dateutil==2.8.2
pdfplumber==0.8.0
PyMuPDF==1.26.3
jinja2==3.1.2
